[data] correct api annotation for tfrecords_datasource
