+++ doc/source/data/api/data_source.rst
@@ -1,0 +1,12 @@
+.. _data-source-api:
+
+Global configuration
+====================
+
+.. currentmodule:: ray.data.datasource
+
+.. autosummary::
+   :nosignatures:
+   :toctree: doc/
+
+   tfrecords_datasource.TFXReadOptions
