+++ python/ray/data/read_api.py
@@ -22,6 +22,7 @@ import ray
 from ray._private.auto_init_hook import wrap_auto_init
 from ray.air.util.tensor_extensions.utils import _create_possibly_ragged_ndarray
 from ray.data._internal.datasource.avro_datasource import AvroDatasource
+from ray.data._internal.datasource.tfrecords_datasource import TFRecordDatasource
 from ray.data._internal.delegating_block_builder import DelegatingBlockBuilder
 from ray.data._internal.logical.operators.from_operators import (
     FromArrow,
@@ -65,7 +66,6 @@ from ray.data.datasource import (
     RangeDatasource,
     SQLDatasource,
     TextDatasource,
-    TFRecordDatasource,
     TorchDatasource,
     WebDatasetDatasource,
 )
